requirements.txt
feedparser
newspaper3k